+++ src/aind_behavior_gym/agent/__init__.py
@@ -1,0 +1,2 @@
+"""Package for `agent` who perform `task` in `gym_env`.
+"""
