+++ src/aind_behavior_gym/task/random_walk_task.py
@@ -4,7 +4,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from aind_behavior_gym.dynamic_foraging_tasks.base import DynamicBanditTask
+from aind_behavior_gym.task.base import DynamicBanditTask
 from aind_behavior_gym.gym_env.dynamic_bandit_env import L, R
