+++ src/aind_behavior_gym/task/uncoupled_block_task.py
@@ -9,7 +9,7 @@ import logging
 import matplotlib.pyplot as plt
 import numpy as np
 
-from aind_behavior_gym.dynamic_foraging_tasks.base import DynamicBanditTask
+from aind_behavior_gym.task.base import DynamicBanditTask
 from aind_behavior_gym.gym_env.dynamic_bandit_env import IGNORE, L, R
 
 logger = logging.getLogger(__name__)
