+++ tests/test_uncoupled_block_task.py
@@ -4,7 +4,7 @@ import unittest
 
 import numpy as np
 
-from aind_behavior_gym.dynamic_foraging_tasks.uncoupled_block_task import UncoupledBlockTask
+from aind_behavior_gym.task.uncoupled_block_task import UncoupledBlockTask
 from aind_behavior_gym.gym_env.dynamic_bandit_env import DynamicBanditEnv, L, R
